Deleted the install instruction image. Read the commit description.
Apparently you can force GitHub to create a cached file url for files you attached to wikis, git issues, and more without having to actual create or save the wiki, issue, and so forth.

Basically we say create new issue and attach the image. It creates a cached git content url  you can copy and put into read me or anywhere else. You don't have to even hit save or create the GitHub issue for it to be cached.