Added a couple new command related classes for characters.
Added:
The CharacterCommandController  class allows for setting up a custom set of commands to be run when invoked.
This allows for creating a list of nodes to be used for custom event running.
Example think of when a character gets hurt you can have a sound effect play, the sprite renderer flicker red, and add some knockback physics to the character.

Also added the start of the status effect node. To allow for setting custom status effects.
An interface for Status Effects is planned to be made soon add more functionality to this.