Moved some properties from GroundedController to it's base class. See Description.
Changed:
This change allows for collision detection changes to invoke events in the Controller2D base class. This is done for characters that might not need gravity, slope detection, and so forth, but still needs some collision detection on the four sides.

Example Thwomp like AI that is grounded, but still needs to check when it hits something below to tell it to go back up.