Created the start of the InteractionController class.
This is useful when wanting to create interaction that require input, collision, or other types of activation for starting an interaction.

Example one: Input interaction to press a button in game to turn on a computer, open a door, or activate items.

Example two: Stepping on a button to activate a pressure plate.