Added a position handle component to help visualize positions needed in other components.
Added:
Position Handle. This script has a custom editor that allows dragging around two position points to allow easily setting Vector2/3 based positions in scene. This can be used to lower the amount of references to game object transforms when needing to set positions for a variety of things.

Example implementation. The Activatable Platform component uses this to allow moving around the start and end position of a moving platform's path.