Created an Activatable Platform component.
Added:
The Activatable Platform component that allows for turning moving platforms on and off when connected to any type of interaction component.

Example usage. The Interactable Switch component allows a switch/button/object to be interacted with to activate anything inheriting from the Activatable wrapper class. This allows for the Interactable Switch to activate a variety of things like Activatable Platforms.