Added the Interactable Switch component.
Added:
The Interactable Switch component allows for setting up a custom interaction to activate any GameObject with any component that inherits from the ActivatableWrapper class. This can be used to turn on/off moving platforms, force fields, teleporters, traps, and a lot more.

At the moment the input version allowing for players to click an input on the keyboard or controllers is fully implemented.
The collision side of interaction is being worked on still.